+++ src/MapperQuery/AbstractMapperQuery.php
@@ -4,12 +4,19 @@ namespace Pyncer\Data\MapperQuery;
 use Pyncer\Data\MapperQuery\MapperQueryInterface;
 use Pyncer\Data\Model\ModelInterface;
 use Pyncer\Database\ConnectionInterface;
+use Pyncer\Database\ConnectionTrait;
 use Pyncer\Database\Record\SelectQueryInterface;
 
 abstract class AbstractMapperQuery implements MapperQueryInterface
 {
+    use ConnectionTrait;
+
+    public function __construct(ConnectionInterface $connection)
+    {
+        $this->setConnection($connection);
+    }
+
     public function overrideModel(
-        ConnectionInterface $connection,
         ModelInterface $model,
         array $data
     ): ModelInterface